Add memoized versions of `rgb_colors` for foreground and background colors.
The changes were made to improve performance by reducing the number of
calculations required to find the nearest color for a given RGB value.
The `rgb_foreground_colors` and `rgb_background_colors` methods only
return colors that match the specified options, which can help reduce
the amount of data that needs to be processed.